requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
chardet==3.0.4
charset-normalizer==3.4.4
deep-translator==1.11.4
feedparser==6.0.12
frozenlist==1.8.0
googletrans==4.0.0rc1
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.26.0
hyperframe==5.2.0
idna==2.10
multidict==6.7.0
propcache==0.4.1
python-telegram-bot==20.8
requests==2.32.5
rfc3986==1.5.0
schedule==1.2.2
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
